MEC172x_AZBEACH: move MC PWM LED driver and fix reset button handling
The MC PWM LED driver is still a board-specific driver for now so move
it under the out_of_tree_boards dir here instead of the main Zephyr
tree.  This will make it easier to update to new Zephyr releases until
the driver is submitted.  Reset button handling is enhanced to allow for
variable control of sending events via SCI or twiddling the SYS_RESET_N
when not in ACPI mode or under driver control.

Signed-off-by: Jeff Daly <jeffd@silicom-usa.com>