MEC172x: Add RPM2PWM block driver
The RPM2PWM block fan control driver is added as a Zephyr PWM driver.
The driver interprets the pulse count of the call to pwm_set_cycles() as
fan RPM vs duty cycle.  The driver includes support for using the tach
reading register as a sensor driver as well.  When instantianting this
device via devicetree a child fan device is required.

Signed-off-by: Jeff Daly <jeffd@silicom-usa.com>